[REFACTOR] extract user model from auth logics
